requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
pydantic
google-generativeai
langchain
faiss-cpu